requirements.txt
astroid==3.3.8
beautifulsoup4==4.12.3
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
dill==0.3.9
distlib==0.3.9
filelock==3.16.1
identify==2.6.6
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mysql-connector-python==9.2.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
pylint==3.3.3
pytest==8.3.4
PyYAML==6.0.2
requests==2.32.3
requests-mock==1.12.1
soupsieve==2.6
tomlkit==0.13.2
urllib3==2.3.0
virtualenv==20.29.1
